% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETBlocos}
\alias{GETBlocos}
\title{Função que retorna Lista de dados sobre os blocos partidários}
\usage{
GETBlocos(
  id = NULL,
  idLegislatura = NULL,
  pagina = NULL,
  itens = NULL,
  ordem = NULL,
  ordenarPor = NULL
)
}
\arguments{
\item{id}{Número(s) identificador(es) de um ou mais bloco(s), separados por vírgulas.}

\item{idLegislatura}{Número(s) identificador(es) de uma ou mais legislatura(s), separados por vírgulas. Se presente, faz com que sejam retornados todos os blocos que existiram nessa(s) legislatura(s).}

\item{pagina}{Número da página de resultados, a partir de 1, que se deseja obter com a requisição, contendo o número de itens definido pelo parâmetro itens. Se omitido, assume o valor 1.}

\item{itens}{Número máximo de itens na página que se deseja obter com esta requisição.}

\item{ordem}{O sentido da ordenação: asc para A a Z ou 0 a 9, e desc para Z a A ou 9 a 0.}

\item{ordenarPor}{Nome do campo pelo qual a lista será ordenada: idLegislatura, id ou nome.}
}
\description{
Nas atividades parlamentares, partidos podem se juntar em blocos partidários. Quando associados, os partidos passam a trabalhar como se fossem um "partidão", com um só líder e um mesmo conjunto de vice-líderes.
Os blocos só podem existir até o fim da legislatura em que foram criados: na legislatura seguinte, os mesmos partidos, se associados, formam um novo bloco.
Este recurso é uma lista dos blocos em atividade no momento da requisição. Se forem passados números de legislaturas com o parâmetro idLegislatura, são listados também os blocos formados e extintos nessas legislaturas.
}
\examples{
GETBlocos()
}
\keyword{blocos}
