% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETBlocosId}
\alias{GETBlocosId}
\title{Função que retorna Informações sobre um bloco partidário específico}
\usage{
GETBlocosId(id)
}
\arguments{
\item{id}{Número identificador do bloco de partidos.}
}
\description{
Retorna informações sobre o bloco identificado por {id} além daquelas obtidas na listagem.
}
\examples{
GETBlocosId(id = '57')
}
\keyword{blocos}
