% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETDeputadoDespesa}
\alias{GETDeputadoDespesa}
\title{Função que retorna As despesas com exercício parlamentar do deputado}
\usage{
GETDeputadoDespesa(
  id,
  idLegislatura = NULL,
  ano = NULL,
  mes = NULL,
  cnpjCpfFornecedor = NULL,
  pagina = NULL,
  itens = NULL,
  ordem = NULL,
  ordenarPor = NULL
)
}
\arguments{
\item{id}{Identificador numérico do parlamentar.}

\item{idLegislatura}{Número(s) de uma ou mais legislatura(s), separados por vírgulas, em que tenham ocorrido as despesas.}

\item{ano}{Um ou mais ano(s) de ocorrência das despesas.}

\item{mes}{Um ou mais número(s) do(s) mês(es) de ocorrência das despesas.}

\item{cnpjCpfFornecedor}{CNPJ de uma pessoa jurídica, ou CPF de uma pessoa física, fornecedora do produto ou serviço (apenas números)}

\item{pagina}{Número da página de resultados, a partir de 1, que se deseja obter com a requisição, contendo o número de itens definido pelo parâmetro itens. Se omitido, assume o valor 1.}

\item{itens}{Número máximo de itens na página que se deseja obter com esta requisição.}

\item{ordem}{O sentido da ordenação: asc para A a Z ou 0 a 9, e desc para Z a A ou 9 a 0.}

\item{ordenarPor}{Nome do campo pelo qual a lista deverá ser ordenada: qualquer um dos campos do retorno, e também idLegislatura}
}
\description{
Dá acesso aos registros de pagamentos e reembolsos feitos pela Câmara em prol do deputado identificado por {id}, a título da Cota para Exercício da Atividade Parlamentar, a chamada "cota parlamentar".
A lista pode ser filtrada por mês, ano, legislatura, CNPJ ou CPF de um fornecedor.
Se não forem passados os parâmetros de tempo, o serviço retorna os dados dos seis meses anteriores à requisição.
}
\examples{
GETDeputadoDespesa(id = '204554')
}
\keyword{deputados}
