% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETDeputadoDiscurso}
\alias{GETDeputadoDiscurso}
\title{Função que retorna Os discursos feitos por um deputado em eventos diversos}
\usage{
GETDeputadoDiscurso(
  id,
  idLegislatura = NULL,
  dataInicio = NULL,
  dataFim = NULL,
  pagina = NULL,
  itens = NULL,
  ordem = NULL,
  ordenarPor = NULL
)
}
\arguments{
\item{id}{Identificador numérico do parlamentar.}

\item{idLegislatura}{Número(s) de uma ou mais legislatura(s), separados por vírgulas, em que tenham ocorrido as despesas.}

\item{dataInicio}{Data de início de um intervalo de tempo, no formato AAAA-MM-DD.}

\item{dataFim}{Data de término de um intervalo de tempo, no formato AAAA-MM-DD.}

\item{pagina}{Número da página de resultados, a partir de 1, que se deseja obter com a requisição, contendo o número de itens definido pelo parâmetro itens. Se omitido, assume o valor 1.}

\item{itens}{Número máximo de itens na página que se deseja obter com esta requisição.}

\item{ordem}{O sentido da ordenação: asc para A a Z ou 0 a 9, e desc para Z a A ou 9 a 0.}

\item{ordenarPor}{Qual dos elementos da representação deverá ser usado para aplicar ordenação à lista.}
}
\description{
Retorna uma lista de informações sobre os pronunciamentos feitos pelo deputado identificado por {id} que tenham sido registrados, em quaisquer eventos, nos sistemas da Câmara.
Caso os parâmetros de tempo (dataInicio, dataFim e idLegislatura) não sejam configurados na requisição, são buscados os discursos ocorridos nos sete dias anteriores ao da requisição
}
\examples{
GETDeputadoDiscurso(id = '204554')
}
\keyword{deputados}
