% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETDeputadoFrentes}
\alias{GETDeputadoFrentes}
\title{Função que retorna As frentes parlamentares das quais um deputado é integrante}
\usage{
GETDeputadoFrentes(id)
}
\arguments{
\item{id}{Identificador numérico do parlamentar.}
}
\description{
Retorna uma lista de informações básicas sobre as frentes parlamentares das quais o parlamentar identificado por {id} seja membro, ou, no caso de frentes existentes em legislaturas anteriores, tenha encerrado a legislatura como integrante.
}
\examples{
GETDeputadoFrentes(id = '204554')
}
\keyword{deputados}
