% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETDeputados}
\alias{GETDeputados}
\title{Função que retorna Listagem e busca de deputados, segundo critérios}
\usage{
GETDeputados(
  id = NULL,
  nome = NULL,
  idLegislatura = NULL,
  siglaUf = NULL,
  siglaPartido = NULL,
  siglaSexo = NULL,
  pagina = NULL,
  itens = NULL,
  dataInicio = NULL,
  dataFim = NULL,
  ordem = NULL,
  ordenarPor = NULL
)
}
\arguments{
\item{id}{Número(s) identificador(es) de um deputado, separados por vírgulas.}

\item{nome}{Parte nome parlamentar.}

\item{idLegislatura}{Número(s) identificador(es) de uma ou mais legislatura(s) de que os parlamentares tenham participado, separados por vírgulas.}

\item{siglaUf}{Uma ou mais sigla(s) de unidades federativas (estados e Distrito Federal). Uma lista de siglas válidas pode ser obtida em /referencias/deputados/siglaUf. Se ausente, serão retornados deputados de todos os estados.}

\item{siglaPartido}{Uma ou mais sigla(s) de partidos aos quais sejam filiados os deputados. Para obter as siglas válidas, consulte /partidos. Atenção: partidos diferentes podem usar a mesma sigla em diferentes legislaturas!}

\item{siglaSexo}{Letra que designe o gênero dos parlamentares que se deseja buscar, sendo M para masculino e F para feminino}

\item{pagina}{Número da página de resultados, a partir de 1, que se deseja obter com a requisição, contendo o número de itens definido pelo parâmetro itens. Se omitido, assume o valor 1.}

\item{itens}{Número máximo de itens na página que se deseja obter com esta requisição.}

\item{dataInicio}{Data de início de um intervalo de tempo, no formato AAAA-MM-DD.}

\item{dataFim}{Data de término de um intervalo de tempo, no formato AAAA-MM-DD.}

\item{ordem}{O sentido da ordenação: asc para A a Z ou 0 a 9, e desc para Z a A ou 9 a 0.}

\item{ordenarPor}{Nome do campo pelo qual a lista deve ser ordenada: id, idLegislatura, nome, siglaUF ou siglaPartido}
}
\description{
Retorna uma lista de dados básicos sobre deputados que estiveram em exercício parlamentar em algum intervalo de tempo.
Se não for passado um parâmetro de tempo, como idLegislatura ou dataInicio, a lista enumerará somente os deputados em exercício no momento da requisição.
}
\examples{
GETDeputados(idLegislatura = '56')
}
\keyword{deputados}
