% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETDeputadosId}
\alias{GETDeputadosId}
\title{Função que retorna Informações detalhadas sobre um deputado específico}
\usage{
GETDeputadosId(id)
}
\arguments{
\item{id}{Identificador numérico do parlamentar.}
}
\description{
Retorna os dados cadastrais de um parlamentar identificado por {id} que, em algum momento da história e por qualquer período, entrou em exercício na Câmara.
}
\examples{
GETDeputados(id = '204554')
}
\keyword{deputados}
