% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETEventoDeputados}
\alias{GETEventoDeputados}
\title{Função que retorna Os deputados participantes de um evento específico}
\usage{
GETEventoDeputados(id)
}
\arguments{
\item{id}{O identificador numérico do evento do qual se deseja informações}
}
\description{
Retorna uma lista de dados resumidos sobre deputados participantes do evento identificado por {id}.
Se o evento já ocorreu, a lista identifica os deputados que efetivamente registraram presença no evento. Se o evento ainda não ocorreu, a lista mostra os deputados que devem participar do evento, por serem convidados ou por serem membros do(s) órgão(s) responsável pelo evento.
}
\examples{
GETEventoDeputados()
}
\keyword{eventos}
