% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETEventoId}
\alias{GETEventoId}
\title{Função que retorna Informações detalhadas sobre um evento específico}
\usage{
GETEventoId(id)
}
\arguments{
\item{id}{O identificador numérico do evento do qual se deseja informações}
}
\description{
Retorna um conjunto detalhado de informações sobre o evento da Câmara identificado por id.
}
\examples{
GETEventoId()
}
\keyword{eventos}
