% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETEventoOrgaos}
\alias{GETEventoOrgaos}
\title{Função que retorna Lista de órgãos organizadores do evento {id}}
\usage{
GETEventoOrgaos(id)
}
\arguments{
\item{id}{O identificador numérico da entidade.}
}
\description{
Retorna uma lista em que cada item é um conjunto mínimo de dados sobre o(s) órgão(s) responsável(veis) pelo evento identificado por {id}. Atualmente, mas provisoriamente, esta informação já vem incorporada ao retorno de /eventos/{id}, mas este endpoint facilita a importação destes dados em planilhas eletrônicas.
}
\examples{
GETEventoOrgaos()
}
\keyword{eventos}
