% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETEventoPauta}
\alias{GETEventoPauta}
\title{Função que retorna Lista de proposições que foram ou deverão ser avaliadas em um evento de caráter deliberativo}
\usage{
GETEventoPauta(id)
}
\arguments{
\item{id}{O identificador numérico da entidade.}
}
\description{
Se o evento {id} for de caráter deliberativo (uma reunião ordinária, por exemplo) este serviço retorna a lista de proposições previstas para avaliação pelos parlamentares. Cada item identifica, se as informações estiverem disponíveis, a proposição avaliada, o regime de preferência para avaliação, o relator e seu parecer, o resultado da apreciação e a votação realizada.
}
\examples{
GETEventoPauta()
}
\keyword{eventos}
