% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETEventos}
\alias{GETEventos}
\title{Função que retorna Lista de eventos ocorridos ou previstos nos diversos órgãos da Câmara}
\usage{
GETEventos(
  id = NULL,
  codTipoEvento = NULL,
  codSituacao = NULL,
  codTipoOrgao = NULL,
  idOrgao = NULL,
  dataInicio = NULL,
  dataFim = NULL,
  horaInicio = NULL,
  horaFim = NULL,
  pagina = NULL,
  itens = NULL,
  ordem = NULL,
  ordenarPor = NULL
)
}
\arguments{
\item{id}{Identificador(es) numérico(s) de um ou mais eventos, separados por vírgulas.}

\item{codTipoEvento}{Um ou mais identificador(es) numérico(s) do(s) tipo(s) de evento que se deseja obter. Os valores válidos podem ser obtidos por uma requisição a /referencias/tiposEvento.}

\item{codSituacao}{Um ou mais identificador(es) numéricos de tipo(s) de situação de evento, separados por vírgula. Valores válidos podem ser obtidos em /referencias/situacoesEvento.}

\item{codTipoOrgao}{Um ou mais identificador(es) numérico(s) de tipo(s) de órgão(s) realizadores dos eventos que se deseja obter, separados por vírgula. Os valores válidos podem ser obtidos em /referencias/tiposOrgao.}

\item{idOrgao}{Um ou mais identificador(es) numérico(s) de órgão(s), separados por vírgula. Os identificadores podem ser obtidos em uma requisição a /orgaos.}

\item{dataInicio}{Data de início de um intervalo de tempo, no formato AAAA-MM-DD.}

\item{dataFim}{Data de término de um intervalo de tempo, no formato AAAA-MM-DD.}

\item{horaInicio}{Hora inicial de um intervalo de tempo, no formato hh:mm, em horário de Brasília.}

\item{horaFim}{Hora final de um intervalo de tempo, no formato hh:mm, em horário de Brasília.}

\item{pagina}{Número da página de resultados, a partir de 1, que se deseja obter com a requisição, contendo o número de itens definido pelo parâmetro itens. Se omitido, assume o valor 1.}

\item{itens}{Número máximo de itens na página que se deseja obter com esta requisição.}

\item{ordem}{O sentido da ordenação: asc para A a Z ou 0 a 9, e desc para Z a A ou 9 a 0.}

\item{ordenarPor}{Nome do campo pelo qual a lista deve ser ordenada: id, dataHoraInicio, dataHoraFim, descricaoSituacao, descricaoTipo ou titulo}
}
\description{
Retorna uma lista cujos elementos trazem informações básicas sobre eventos dos órgãos legislativos da Câmara, previstos ou já ocorridos, em um certo intervalo de tempo.
Esse intervalo pode ser configurado pelos parâmetros de data e hora listados abaixo. Se nenhum for passado, são listados eventos dos cinco dias anteriores, dos cinco dias seguintes e do próprio dia em que é feita a requisição.
}
\examples{
GETEventos()
}
\keyword{eventos}
