% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETEventoVotacoes}
\alias{GETEventoVotacoes}
\title{Função que retorna Informações detalhadas de votações sobre um evento específico}
\usage{
GETEventoVotacoes(id)
}
\arguments{
\item{id}{O identificador numérico da entidade.}
}
\description{
Retorna uma lista de dados básicos sobre votações que tenham sido realizadas no evento identificado por {id}. Votações só ocorrem em eventos de caráter deliberativo. Dados complementares sobre cada votação listada podem ser obtidos no recurso /votacoes/{id}.
Para compreender melhor os dados sobre votações, veja a página de tutorial do Portal de Dados Abertos.
}
\examples{
GETEventoVotacoes()
}
\keyword{eventos}
