% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETFrenteId}
\alias{GETFrenteId}
\title{Função que retorna Informações detalhadas sobre uma frente parlamentar}
\usage{
GETFrenteId(id)
}
\arguments{
\item{id}{Identificador numérico da frente parlamentar.}
}
\description{
Este recurso traz informações detalhadas sobre a frente parlamentar identificada por {id}.
}
\examples{
GETFrenteId()
}
\keyword{frentes}
