% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETFrenteMembros}
\alias{GETFrenteMembros}
\title{Função que retorna Os deputados que participam de uma frente parlamentar}
\usage{
GETFrenteMembros(id)
}
\arguments{
\item{id}{Identificador numérico da frente parlamentar.}
}
\description{
Uma lista dos deputados participantes da frente parlamentar identificada por {id} e os papéis que exerceram nessa frente (signatário, coordenador ou presidente).
Observe que, mesmo no caso de frentes parlamentares mistas (compostas por deputados e senadores), são retornados apenas dados sobre os deputados.
}
\examples{
GETFrenteMembros()
}
\keyword{frentes}
