% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETFrentes}
\alias{GETFrentes}
\title{Função que retorna Lista de frentes parlamentares de uma ou mais legislaturas}
\usage{
GETFrentes(idLegislatura = NULL, pagina = NULL)
}
\arguments{
\item{idLegislatura}{Número da(s) legislatura(s), separados por vírgulas, às quais os dados buscados devem corresponder.}

\item{pagina}{Número da página de resultados, a partir de 1, que se deseja obter com a requisição, contendo o número de itens definido pelo parâmetro itens. Se omitido, assume o valor 1.}
}
\description{
Retorna uma lista de informações sobre uma frente parlamentar - um agrupamento oficial de parlamentares em torno de um determinado tema ou proposta.
As frentes existem até o fim da legislatura em que foram criadas, e podem ser recriadas a cada legislatura. Algumas delas são compostas por deputados e senadores.
Um ou mais número de legislatura(s) pode(m) ser passado(s) como parâmetro, mas se for omitido são retornadas todas as frentes parlamentares criadas desde 2003.
}
\examples{
GETFrentes()
}
\keyword{frentes}
