% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETLegislatura}
\alias{GETLegislatura}
\title{Função que retorna Os períodos de mandatos e atividades parlamentares da Câmara}
\usage{
GETLegislatura(
  id = NULL,
  data = NULL,
  pagina = NULL,
  itens = NULL,
  ordem = NULL,
  ordenarPor = NULL
)
}
\arguments{
\item{id}{Um ou mais número(s) de legislatura(s), separados por vírgulas. Se omitido, serão retornados dados sobre todas as legislaturas.}

\item{data}{Uma ou mais data(s) no formato AAAA-MM-DD, separadas por vírgulas. Se este parâmetro estiver presente, a requisição retornará as informações básicas sobre as legislaturas que estavam em curso nas datas informadas.}

\item{pagina}{Número da página de resultados, a partir de 1, que se deseja obter com a requisição, contendo o número de itens definido pelo parâmetro itens. Se omitido, assume o valor 1.}

\item{itens}{Número máximo de itens na página que se deseja obter com esta requisição.}

\item{ordem}{O sentido da ordenação: asc para A a Z ou 0 a 9, e desc para Z a A ou 9 a 0.}

\item{ordenarPor}{Qual dos elementos da representação deverá ser usado para aplicar ordenação à lista.}
}
\description{
Legislatura é o nome dado ao período de trabalhos parlamentares entre uma eleição e outra.
Este serviço retorna uma lista em que cada item contém as informações básicas sobre um desses períodos.
Os números que identificam as legislaturas são sequenciais, desde a primeira que ocorreu.
}
\examples{
GETLegislatura(id = '56')
}
\keyword{LEGISLATURAS}
