% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETLegislaturaId}
\alias{GETLegislaturaId}
\title{Função que retorna Informações extras sobre uma determinada legislatura da Câmara}
\usage{
GETLegislaturaId(id)
}
\arguments{
\item{id}{Número da legislatura da qual se quer os dados.}
}
\description{
Retorna informações adicionais sobre o período de atividades da Câmara identificado por {id}.
}
\examples{
GETLegislaturaId(id = '56')
}
\keyword{LEGISLATURAS}
