% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETLegislaturaMesa}
\alias{GETLegislaturaMesa}
\title{Função que retorna Quais deputados fizeram parte da Mesa Diretora em uma legislatura}
\usage{
GETLegislaturaMesa(idLegislatura, dataInicio = NULL, dataFim = NULL)
}
\arguments{
\item{dataInicio}{Dia de início do intervalo de tempo do qual se deseja saber a composição da Mesa, no formato AAAA-MM-DD.}

\item{dataFim}{Data de término do intervalo de tempo do qual se deseja saber a composição da Mesa, no formato AAAA-MM-DD.}

\item{id}{Número da legislatura da qual se deseja os dados.}
}
\description{
Retorna uma lista com dados básicos sobre todos os deputados que ocuparam algum posto na Mesa Diretora da Câmara em algum período de tempo dentro da legislatura identificada por {id}.
Normalmente, cada legislatura tem duas Mesas Diretoras, com presidente, dois vice-presidentes, quatro secretários parlamentares e os suplentes dos secretários.
}
\examples{
GETLegislaturaMesa(id = '56')
}
\keyword{LEGISLATURAS}
