% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETOrgaoMembros}
\alias{GETOrgaoMembros}
\title{Função que retorna Lista de cargos de um órgão e parlamentares que os ocupam.}
\usage{
GETOrgaoMembros(
  id,
  dataInicio = NULL,
  dataFim = NULL,
  pagina = NULL,
  itens = NULL
)
}
\arguments{
\item{id}{Identificador numérico do órgão}

\item{dataInicio}{Data de início de um intervalo de tempo, no formato AAAA-MM-DD.}

\item{dataFim}{Data de término de um intervalo de tempo, no formato AAAA-MM-DD.}

\item{pagina}{Número da página de resultados, a partir de 1, que se deseja obter com a requisição, contendo o número de itens definido pelo parâmetro itens. Se omitido, assume o valor 1.}

\item{itens}{Número máximo de itens na página que se deseja obter com esta requisição.}
}
\description{
Retorna uma lista de dados resumidos que identificam cada parlamentar e o cargo ou posição que ocupa ou ocupou no órgão parlamentar identificado por {id} durante um certo período de tempo.
Se não forem passados parâmetros que delimitem esse período, o serviço retorna os membros do órgão no momento da requisição. Se o órgão não existir mais ou não estiver instalado, é retornada uma lista vazia.
}
\examples{
GETOrgaoMembros()
}
\keyword{orgãos}
