% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETOrgaos}
\alias{GETOrgaos}
\title{Função que retorna A lista das comissões e outros órgãos legislativos da Câmara}
\usage{
GETOrgaos(
  id = NULL,
  sigla = NULL,
  codTipoOrgao = NULL,
  dataInicio = NULL,
  dataFim = NULL,
  pagina = NULL,
  itens = NULL,
  ordem = NULL,
  ordenarPor = NULL
)
}
\arguments{
\item{id}{Um ou mais identificador(es) numérico(s), separados por vírgulas.}

\item{sigla}{Uma ou mais sigla(s) oficialmente usadas para designar um ou mais órgão(s) da Câmara, separadas por vírgulas.}

\item{codTipoOrgao}{Um ou mais identificador(es) numérico(s) do(s) tipo(s) de órgãos que se deseja buscar dados. Pode(m) ser obtido(s) em /referencias/orgaos/codTipoOrgao.}

\item{dataInicio}{Data de início, no formato AAAA-MM-DD, de um intervalo de tempo no qual os órgãos buscados devem ter estado em atividade.}

\item{dataFim}{Data de término, no formato AAAA-MM-DD, de um intervalo de tempo no qual os órgãos buscados devem ter estado em atividade.}

\item{pagina}{Número da página de resultados, a partir de 1, que se deseja obter com a requisição, contendo o número de itens definido pelo parâmetro itens. Se omitido, assume o valor 1.}

\item{itens}{Número máximo de itens na página que se deseja obter com esta requisição.}

\item{ordem}{O sentido da ordenação: asc para A a Z ou 0 a 9, e desc para Z a A ou 9 a 0.}

\item{ordenarPor}{Nome do campo pelo qual a lista será ordenada: id, sigla, nome, apelido, codTipoOrgao, tipoOrgao, dataInicio ou dataFim}
}
\description{
Retorna uma lista de informações básicas sobre os órgãos legislativos e seus identificadores, tipos e descrições.
Pela query string é possível filtrar a lista por identificadores, tipos de órgãos, sigla, situação do órgão ou período de tempo em que os órgãos estiveram ativos, se aplicável.
}
\examples{
GETOrgaos()
}
\keyword{orgãos}
