% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETOrgaosId}
\alias{GETOrgaosId}
\title{Função que retorna Informações detalhadas sobre um órgão da Câmara}
\usage{
GETOrgaosId(id)
}
\arguments{
\item{id}{Identificador numérico do órgão}
}
\description{
Retorna todas as informações disponíveis sobre o órgão da Câmara identificado por id.
}
\examples{
GETOrgaosId()
}
\keyword{orgãos}
