% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETOrgaoVotacoes}
\alias{GETOrgaoVotacoes}
\title{Função que retorna Informações detalhadas sobre votações de um órgão da Câmara}
\usage{
GETOrgaoVotacoes(
  id,
  idProposicao = NULL,
  dataInicio = NULL,
  dataFim = NULL,
  pagina = NULL,
  itens = NULL,
  ordem = NULL,
  ordenarPor = NULL
)
}
\arguments{
\item{id}{Identificador numérico do órgão}

\item{idProposicao}{Um ou mais identificador(es) numéricos de proposições, que podem ser obtidos por meio do recurso /proposicoes. Se presente, listará as votações que tiveram a(s) proposição(ções) como objeto de votação ou que afetaram as proposições listadas.}

\item{dataInicio}{Data em formato AAAA-MM-DD para início do intervalo de tempo no qual tenham sido realizadas as votações a serem listadas. Se usado sozinho, esse parâmetro faz com que sejam retornadas votações ocorridas dessa data até o fim do mesmo ano. Se usado com dataFim, as duas datas devem ser de um mesmo ano.}

\item{dataFim}{Data em formato AAAA-MM-DD que define o fim do intervalo de tempo no qual tenham sido realizadas as votações a serem listadas. Se usado sozinho, esse parâmetro faz com que sejam retornadas todas as votações ocorridas desde 1º de janeiro do mesmo ano até esta data. Se usado com dataInicio, é preciso que as duas datas sejam de um mesmo ano.}

\item{pagina}{Número da página de resultados, a partir de 1, que se deseja obter com a requisição, contendo o número de itens definido pelo parâmetro itens. Se omitido, assume o valor 1.}

\item{itens}{Número máximo de itens na página que se deseja obter com esta requisição. O valor padrão e máximo para este endpoint é 200, e valores maiores serão ignorados.}

\item{ordem}{O sentido da ordenação: asc para A a Z ou 0 a 9, e desc para Z a A ou 9 a 0.}

\item{ordenarPor}{Nome do campo pelo qual a lista será ordenada. Pode ser id, idOrgao, siglaOrgao, idEvento, idProposicao, data, dataHoraRegistro ou idProposicaoObjeto.}
}
\description{
Retorna uma lista de dados básicos de votações que tenham sido realizadas em eventos realizados no órgão {id}.
Se {id} for um órgão permanente da Câmara, são retornados, por padrão, dados sobre as votações realizadas pelo órgão nos últimos 30 dias. Esse período pode ser alterado com o uso dos parâmetros dataInicio e/ou dataFim, que por enquanto são limitados a selecionar somente votações ocorridas em um mesmo ano.
Caso {id} seja um órgão temporário, como uma comissão especial, são listadas por padrão todas as votações ocorridas no órgão, em qualquer período de tempo.
Dados complementares sobre cada votação listada podem ser obtidos no recurso /votacoes/{id}.
Para compreender melhor os dados sobre votações, veja a página de tutorial do Portal de Dados Abertos.
}
\examples{
GETOrgaoVotacoes()
}
\keyword{orgãos}
