% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETPartidoId}
\alias{GETPartidoId}
\title{Função que retorna Informações detalhadas sobre um partido}
\usage{
GETPartidoId(id)
}
\arguments{
\item{id}{Id do Partido}
}
\description{
Função que retorna Informações detalhadas sobre um partido
}
\examples{
GETPartidoId()
}
\keyword{partidos}
