% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETPartidos}
\alias{GETPartidos}
\title{Função que retorna Os partidos políticos que têm ou já tiveram parlamentares em exercício na Câmara}
\usage{
GETPartidos(
  sigla = NULL,
  dataInicio = NULL,
  dataFim = NULL,
  idLegislatura = NULL,
  pagina = NULL,
  itens = NULL,
  ordem = NULL,
  ordenarPor = NULL
)
}
\arguments{
\item{sigla}{Sigla de um ou mais partido(s), separadas por vírgulas.}

\item{dataInicio}{Data de início de um intervalo de tempo, no formato AAAA-MM-DD.}

\item{dataFim}{Data de término de um intervalo de tempo, no formato AAAA-MM-DD.}

\item{idLegislatura}{Número da(s) legislatura(s), separados por vírgulas, às quais os dados buscados devem corresponder.}

\item{pagina}{Número da página de resultados, a partir de 1, que se deseja obter com a requisição, contendo o número de itens definido pelo parâmetro itens. Se omitido, assume o valor 1.}

\item{itens}{Número máximo de itens na página que se deseja obter com esta requisição.}

\item{ordem}{O sentido da ordenação: asc para A a Z ou 0 a 9, e desc para Z a A ou 9 a 0.}

\item{ordenarPor}{Nome do campo pelo qual a lista deve ser ordenada: id, sigla, nome, dataInicio ou dataFim}
}
\description{
Retorna uma lista de dados básicos sobre os partidos políticos que têm ou já tiveram deputados na Câmara.
Se não forem passados parâmetros, o serviço retorna os partidos que têm deputados em exercício no momento da requisição.
É possível obter uma lista de partidos representados na Câmara em um certo intervalo de datas ou de legislaturas. Se um intervalo e uma ou mais legislatura(s) não coincidentes forem passados, todos os intervalos de tempo serão somados.
Também se pode fazer busca por uma ou mais sigla(s), mas atenção: em diferentes legislaturas, pode haver mais de um partido usando a mesma sigla.
}
\examples{
GETPartidos(sigla = 'PSL')
}
\keyword{partidos}
