% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETProposicaoId}
\alias{GETProposicaoId}
\title{Função que retorna Informações detalhadas sobre uma proposição específica}
\usage{
GETProposicaoId(id)
}
\arguments{
\item{id}{Identificador numérico da proposição}
}
\description{
Função que retorna Informações detalhadas sobre uma proposição específica
}
\examples{
GETProposicaoId()
}
\keyword{proposições}
