% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETProposicaoRelacionadas}
\alias{GETProposicaoRelacionadas}
\title{Função que retorna Uma lista de proposições relacionadas a uma em especial}
\usage{
GETProposicaoRelacionadas(id)
}
\arguments{
\item{id}{Identificador numérico da proposição}
}
\description{
Lista de informações básicas sobre proposições que de alguma forma se relacionam com a proposição identificada por {id}, como pareceres, requerimentos, substitutivos, etc.
}
\examples{
GETProposicaoRelacionadas()
}
\keyword{proposições}
