% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETProposicaoTemas}
\alias{GETProposicaoTemas}
\title{Função que retorna Lista de áreas temáticas de uma proposição}
\usage{
GETProposicaoTemas(id)
}
\arguments{
\item{id}{Identificador numérico da proposição}
}
\description{
Lista em que cada item traz informações sobre uma área temática à qual a proposição identificada por {id} se relaciona, segundo classificação oficial do Centro de Documentação e Informação da Câmara.
}
\examples{
GETProposicaoTemas()
}
\keyword{proposições}
