% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETProposicaoTramitacoes}
\alias{GETProposicaoTramitacoes}
\title{Função que retorna O histórico de passos na tramitação de uma proposta}
\usage{
GETProposicaoTramitacoes(id, dataInicio = NULL, dataFim = NULL)
}
\arguments{
\item{id}{Identificador numérico da proposição}

\item{dataInicio}{Data do início da tramitação, no formato AAAA-MM-DD.}

\item{dataFim}{Data do fim da tramitação, no formato AAAA-MM-DD.}
}
\description{
Lista que traz, como cada item, um “retrato” de informações que podem ser alteradas a cada etapa de tramitação na vida de uma proposição (como regime de tramitação e situação) e informações sobre o que causou esse novo estado.
Esta representação das tramitações ainda é provisória.
}
\examples{
GETProposicaoTramitacoes()
}
\keyword{proposições}
