% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETProposicaoVotacoes}
\alias{GETProposicaoVotacoes}
\title{Função que retorna Informações detalhadas de votações sobre uma proposição específica}
\usage{
GETProposicaoVotacoes(id, ordem = NULL, ordenarPor = NULL)
}
\arguments{
\item{id}{Identificador numérico da proposição}

\item{ordem}{Nome do campo pelo qual a lista será ordenada: id, dataHoraRegistro}

\item{ordenarPor}{Identificador numérico da proposição}
}
\description{
Retorna uma lista de identificadores básicos sobre as votações na Câmara que tiveram a proposição {id} como objeto ou como afetada pelos seus resultados. Dados complementares sobre cada votação listada podem ser obtidos no recurso /votacoes/{id}.
Para compreender melhor os dados sobre votações, veja a página de tutorial do Portal de Dados Abertos.
}
\examples{
GETProposicaoVotacoes()
}
\keyword{proposições}
