% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETProposicoes}
\alias{GETProposicoes}
\title{Função que retorna Lista configurável de proposições na Câmara}
\usage{
GETProposicoes(
  id = NULL,
  siglaTipo = NULL,
  numero = NULL,
  ano = NULL,
  idDeputadoAutor = NULL,
  autor = NULL,
  siglaPartidoAutor = NULL,
  idPartidoAutor = NULL,
  siglaUfAutor = NULL,
  keywords = NULL,
  tramitacaoSenado = NULL,
  dataInicio = NULL,
  dataFim = NULL,
  dataApresentacaoInicio = NULL,
  dataApresentacaoFim = NULL,
  codSituacao = NULL,
  codTema = NULL,
  pagina = NULL,
  itens = NULL,
  ordem = NULL,
  ordenarPor = NULL
)
}
\arguments{
\item{id}{Número(s) identificador(es) de uma ou mais proposições no Dados Abertos, separados por vírgulas.}

\item{siglaTipo}{Uma ou mais sigla(s) separadas por vírgulas do(s) tipo(s) das proposições que se deseja obter. A lista de tipos e siglas existentes pode ser obtida em /referencias/proposicoes/siglaTipo}

\item{numero}{Um ou mais número(s), separados por vírgula, oficialmente atribuídos às proposições segundo o art. 137 do Regimento Interno, como “PL 1234/2016”}

\item{ano}{Um ou mais ano(s) de apresentação das proposições que serão listadas, separados por vírgulas, no formato AAAA}

\item{idDeputadoAutor}{Um ou mais números identificador(es), separados por vírgula, do(s) deputado(s) autor(es) das proposições que serão listadas. Cada número deve ser o identificador exclusivo de um parlamentar no Dados Abertos}

\item{autor}{Nome ou parte do nome do(s) autor(es) das proposições que se deseja obter. Deve estar entre aspas}

\item{siglaPartidoAutor}{Uma ou mais sigla(s) separadas por vírgulas do(s) partido(s) a que pertençam os autores das proposições a serem listadas}

\item{idPartidoAutor}{Identificador numérico no Dados Abertos do partido a que pertençam os autores das proposições que serão listadas. Esses identificadores podem ser obtidos em /partidos e são mais precisos do que as siglas, que podem ser usadas por partidos diferentes em épocas diferentes}

\item{siglaUfAutor}{Uma ou mais sigla(s) de unidade(s) da federação (estados e Distrito Federal) pela(s) qual(quais) o(s) autor(es) das proposições selecionadas tenha(m) sido eleito(s)}

\item{keywords}{Uma ou mais palavras chaves sobre o tema a que a proposição se relaciona}

\item{tramitacaoSenado}{Indicador booleano, com valor TRUE ou FALSE para trazer apenas proposições que já tenha tramitado no Senado}

\item{dataInicio}{Data do início do intervalo de tempo em que tenha havido tramitação das proposições a serem listadas, no formato AAAA-MM-DD. Se omitido, é assumido como a data de 30 dias anteriores à proposição}

\item{dataFim}{Data do fim do intervalo de tempo em que tenha havido tramitação das proposições a serem listadas. Se omitido, é considerado ser o dia em que é feita a requisição}

\item{dataApresentacaoInicio}{Data do início do intervalo de tempo em que tenham sido apresentadas as proposições a serem listadas, no formato AAAA-MM-DD}

\item{dataApresentacaoFim}{Data do fim do intervalo de tempo em que tenham sido apresentadas as proposições a serem listadas}

\item{codSituacao}{Código(s) numérico(s), separados por vírgulas, do tipo de situação em que se encontram as proposições que serão listadas. As situações possíveis podem ser obtidas em /referencias/proposicoes/codSituacao. Atenção: este parâmetro pode apresentar resultados inesperados, por problemas com o registro dos dados.}

\item{codTema}{Código(s) numérico(s), separados por vírgulas, das áreas temáticas das proposições que serão listadas. Os temas possíveis podem ser obtidos em /referencias/proposicoes/codTema}

\item{pagina}{Número da página de resultados, a partir de 1, que se deseja obter com a requisição, contendo o número de itens definido pelo parâmetro itens. Se omitido, assume o valor 1.}

\item{itens}{Número máximo de itens na página que se deseja obter com esta requisição.}

\item{ordem}{O sentido da ordenação: asc para A a Z ou 0 a 9, e desc para Z a A ou 9 a 0.}

\item{ordenarPor}{Nome do campo pelo qual a lista deve ser ordenada: id, codTipo, siglaTipo, numero ou ano}
}
\description{
Lista de informações básicas sobre projetos de lei, resoluções, medidas provisórias, emendas, pareceres e todos os outros tipos de proposições na Câmara.
Por padrão, são retornadas todas as proposições que foram apresentadas ou tiveram alguma mudança de situação nos últimos 30 dias. Esse intervalo de tramitação pode ser configurado pelos parâmetros dataInicio e dataFim.
Se for(em) passado(s) um ou mais dos parâmetros…
id
ano
dataApresentacaoInicio
dataApresentacaoFim
idAutor
autor
…o intervalo de tramitação só será levado em consideração se os parâmetros dataInicio e/ou dataFim estiverem explicitamente configurados. Se não estiverem, poderão ser listadas proposições que não tiveram tramitação recente (e a resposta pode demorar bastante).
}
\examples{
GETProposicoes( ano = '2020' )
}
\keyword{proposições}
