% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETVotacaoOrientacao}
\alias{GETVotacaoOrientacao}
\title{Função que retorna O voto recomendado pelas lideranças aos seus deputados em uma votação}
\usage{
GETVotacaoOrientacao(id)
}
\arguments{
\item{id}{Identificador alfanumérico da votação}
}
\description{
Em muitas votações, os líderes de partidos e blocos – as bancadas – fazem recomendações de voto para seus parlamentares. Essas orientações de uma votação também são feitas pelas lideranças de Governo, Minoria e as mais recentes Maioria e Oposição. Uma liderança também pode liberar a bancada para que cada deputado vote como quiser, ou entrar em obstrução, para que seus parlamentares não sejam contados para o quórum da votação.
Se a votação identificada por {id} teve orientações, este recurso retorna uma lista em que cada item contém os identificadores de um partido, bloco ou liderança, e o posicionamento ou voto que foi recomendado aos seus parlamentares.
Até o momento, só estão disponíveis dados sobre orientações dadas em votações no Plenário.
}
\examples{
GETVotacaoOrientacao()
}
\keyword{votações}
