% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETVotacaoVotos}
\alias{GETVotacaoVotos}
\title{Função que retorna Como cada parlamentar votou em uma votação nominal e aberta}
\usage{
GETVotacaoVotos(id)
}
\arguments{
\item{id}{Identificador alfanumérico da votação}
}
\description{
Se {id} é o identificador de uma votação da Câmara nominal que não tenha sido secreta, este endpoint retorna uma lista em que cada item contém os identificadores básicos de um deputado e o voto ou posicionamento que ele registrou.
O resultado é uma lista vazia se {id} foi uma votação simbólica, em que os votos individuais não são contabilizados. Mas há algumas votações simbólicas que também têm registros de "votos": nesses casos, normalmente se trata de parlamentares que pediram expressamente que seus posicionamentos fossem registrados.
Não são listados parlamentares ausentes à votação.
}
\examples{
GETVotacaoVotos()
}
\keyword{votações}
