% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GETVotacoes}
\alias{GETVotacoes}
\title{Função que retorna Lista das votações da Câmara}
\usage{
GETVotacoes(
  id = NULL,
  idProposicao = NULL,
  idEvento = NULL,
  idOrgao = NULL,
  dataInicio = NULL,
  dataFim = NULL,
  pagina = NULL,
  itens = NULL,
  ordem = NULL,
  ordenarPor = NULL
)
}
\arguments{
\item{id}{Um ou mais identificador(es) alfanuméricos de votação, separados por vírgulas, para que seja(m) listado(s) dados sobre uma ou mais votações específicas.}

\item{idProposicao}{Um ou mais identificador(es) numéricos de proposições, que podem ser obtidos por meio do recurso /proposicoes. Se presente, listará as votações que tiveram a(s) proposição(ções) como objeto de votação ou que afetaram as proposições listadas.}

\item{idEvento}{Identificador de um ou mais evento(s) realizado(s) na Câmara, separados por vírgula, nos quais tenham sido realizadas as votações a serem listadas. Os identificadores podem ser obtidos por meio do recurso /eventos. Somente os eventos deliberativos podem ter votações. Os eventos podem ter ocorrido fora do intervalo de tempo padrão ou definido por dataInicio e/ou dataFim.}

\item{idOrgao}{Um ou mais identificador(es) numéricos de órgãos da Câmara, separados por vírgulas. Se presente, serão retornadas somente votações dos órgãos enumerados. Os identificadores existentes podem ser obtidos por meio do recurso /orgaos.}

\item{dataInicio}{Data em formato AAAA-MM-DD para início do intervalo de tempo no qual tenham sido realizadas as votações a serem listadas. Se usado sozinho, esse parâmetro faz com que sejam retornadas votações ocorridas dessa data até o fim do mesmo ano. Se usado com dataFim, as duas datas devem ser de um mesmo ano.}

\item{dataFim}{Data em formato AAAA-MM-DD que define o fim do intervalo de tempo no qual tenham sido realizadas as votações a serem listadas. Se usado sozinho, esse parâmetro faz com que sejam retornadas todas as votações ocorridas desde 1º de janeiro do mesmo ano até esta data. Se usado com dataInicio, é preciso que as duas datas sejam de um mesmo ano.}

\item{pagina}{Número da página de resultados, a partir de 1, que se deseja obter com a requisição, contendo o número de itens definido pelo parâmetro itens. Se omitido, assume o valor 1.}

\item{itens}{Número máximo de itens na página que se deseja obter com esta requisição. O valor padrão e máximo para este endpoint é 200, e valores maiores serão ignorados.}

\item{ordem}{O sentido da ordenação: asc para A a Z ou 0 a 9, e desc para Z a A ou 9 a 0.}

\item{ordenarPor}{Nome do campo pelo qual a lista será ordenada. Pode ser id, idOrgao, siglaOrgao, idEvento, idProposicao, data, dataHoraRegistro ou idProposicaoObjeto.}
}
\description{
Retorna uma lista de informações básicas sobre as votações ocorridas em eventos dos diversos órgãos da Câmara.
Se não forem passados parâmetros que delimitem o intervalo de tempo da pesquisa, são retornados dados sobre todas as votações ocorridas nos últimos 30 dias, em eventos de todos os órgãos.
Os parâmetros de data permitem estender o período, mas por enquanto é necessário que as duas datas sejam de um mesmo ano. Quando apenas uma delas está presente, são retornadas somente as votações ocorridas no mesmo ano, antes de dataFim ou após dataInicio.
Também é possível filtrar a listagem por identificadores de órgãos da Câmara, de proposições e de eventos.
Quando não há identificação da proposição que foi efetivamente votada, é preciso consultar o endpoint /votacoes/{id} para obter uma lista de proposições das quais uma pode ter sido o objeto da votação.
Para mais informações sobre o uso dos endpoints de votações, veja a página de tutorial do Portal de Dados Abertos.
}
\examples{
GETVotacoes()
}
\keyword{votações}
